emb/requirements.txt
pgvector==0.2.5
requests>=2.0.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-decouple==3.8
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
sentence-transformers==2.2.2
transformers==4.35.2
fastapi==0.109.2
uvicorn==0.27.1